Drastic reorg of where Blob, Commit, Tag and Tree are defined
which had many repercussions!
