+++ src/Git/Objects.hs
@@ -1,4 +1,10 @@
 module Git.Objects (module X) where
 
+import Git.Objects.Blob as X
+import Git.Objects.Commit as X
+import Git.Objects.GitObject as X
 import Git.Objects.Object as X
+import Git.Objects.Serialise as X
+import Git.Objects.Tag as X
+import Git.Objects.Tree as X
 import Git.Objects.Types as X
