+++ src/Git/Objects/Tag.hs
@@ -1,0 +1,11 @@
+module Git.Objects.Tag where
+
+import Git.Objects.GitObject (GitObject(..), ObjectType(..))
+
+
+data Tag = Tag
+
+instance GitObject Tag where
+  gitObjectType _ = ObjTyTag
+  encodeObject = undefined
+  objectParser = undefined
