+++ src/Git/Pack/Pack.hs
@@ -24,10 +24,9 @@ import System.IO.MMap (mmapFileForeignPtr, Mode(..))
 import Text.Printf (printf)
 
 import Git.Internal
-  (MmapHandle, mmapData, MmapFrom(..), MmapTo(..), mmapSha1)
+  ( MmapHandle, mmapData, MmapFrom(..), MmapTo(..), mmapSha1
+  , tellParsePos, lazyParseOnly)
 import Git.Objects (ObjectType(..))
--- FIXME: tellParsePos, lazyParseOnly should live somewhere more general
-import Git.Objects.Serialise (tellParsePos, lazyParseOnly)
 import Git.Sha1 (Sha1, sha1Size)
 import qualified Git.Sha1 as Sha1
 import Git.Types (GitError(..))
